--- Feature.java
@@ -1,8 +1,0 @@
-package lab3;
-
-public enum Feature {
-    FLAT,
-    UNEVEN,
-    NARROW,
-    WIDE
-}
