--- GrassLandNotFlatException.java
@@ -1,7 +1,0 @@
-package lab3;
-
-public class GrassLandNotFlatException extends RuntimeException{
-    GrassLandNotFlatException(String mes){
-        super(mes);
-    }
-}
