--- River.java
@@ -1,11 +1,0 @@
-package lab3;
-
-public class River extends Place{
-    public River(String name){
-        super(name);
-    }
-
-    public void Flow(){
-        System.out.print("River turned and tumbled merrily among the high rocky banks\n");
-    }
-}
