--- Say.java
@@ -1,5 +1,0 @@
-package lab3;
-
-public interface Say {
-    void say(String words);
-}
