--- Sex.java
@@ -1,6 +1,0 @@
-package lab3;
-
-public enum Sex {
-    MALE,
-    FEMALE
-}
