--- Status.java
@@ -1,8 +1,0 @@
-package lab3;
-
-public enum Status {
-    WALKING,
-    EATING,
-    ESTIMATE,
-    SITTING
-}
