--- Walkingto.java
@@ -1,5 +1,0 @@
-package lab3;
-
-public interface Walkingto {
-    void walkto(Place desname);
-}
