--- Walkthrough.java
@@ -1,5 +1,0 @@
-package lab3;
-
-public interface Walkthrough {
-    void walkthrough(Place path);
-}
