+++ src/Feature.java
@@ -1,0 +1,6 @@
+public enum Feature {
+    FLAT,
+    UNEVEN,
+    NARROW,
+    WIDE
+}
