+++ src/GrassLandNotFlatException.java
@@ -1,0 +1,5 @@
+public class GrassLandNotFlatException extends RuntimeException{
+    GrassLandNotFlatException(String mes){
+        super(mes);
+    }
+}
