+++ src/Say.java
@@ -1,0 +1,3 @@
+public interface Say {
+    void say(String words);
+}
