+++ src/Sex.java
@@ -1,0 +1,4 @@
+public enum Sex {
+    MALE,
+    FEMALE
+}
