+++ src/Status.java
@@ -1,0 +1,6 @@
+public enum Status {
+    WALKING,
+    EATING,
+    ESTIMATE,
+    SITTING
+}
