+++ src/Walkingto.java
@@ -1,0 +1,3 @@
+public interface Walkingto {
+    void walkto(Place desname);
+}
