+++ src/Walkthrough.java
@@ -1,0 +1,3 @@
+public interface Walkthrough {
+    void walkthrough(Place path);
+}
